# api/AlloBoulot/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation de l'API Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/job-applications": {
            "post": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Postuler à une offre d'emploi",
                "description": "Permet à un utilisateur de postuler à une offre d'emploi avec son CV et lettre de motivation",
                "operationId": "f3347c0843b47753de8374660d410d29",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "job_offer_id",
                                    "resume_ref"
                                ],
                                "properties": {
                                    "job_offer_id": {
                                        "description": "ID de l'offre d'emploi",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Je suis très intéressé par ce poste..."
                                    },
                                    "resume_ref": {
                                        "description": "CV (PDF, DOCX, max 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "motivation_ref": {
                                        "description": "Lettre de motivation (PDF, DOCX, max 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidature envoyée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Candidature envoyée avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "job_offer_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Je suis très intéressé..."
                                                },
                                                "resume_ref": {
                                                    "type": "string",
                                                    "example": "/storage/resumes/cv_123.pdf"
                                                },
                                                "motivation_ref": {
                                                    "type": "string",
                                                    "example": "/storage/motivations/lettre_123.pdf"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/job-applications/offer/{offerId}": {
            "get": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Liste des candidatures pour une offre",
                "description": "Récupère toutes les candidatures pour une offre d'emploi spécifique (RH/Admin uniquement)",
                "operationId": "978f50f779f16119d9e75646d570adb9",
                "parameters": [
                    {
                        "name": "offerId",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des candidatures récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "job_offer_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "Je suis très intéressé..."
                                                    },
                                                    "resume_ref": {
                                                        "type": "string",
                                                        "example": "/storage/resumes/cv_123.pdf"
                                                    },
                                                    "motivation_ref": {
                                                        "type": "string",
                                                        "example": "/storage/motivations/lettre_123.pdf"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "first_name": {
                                                                "type": "string",
                                                                "example": "Jean"
                                                            },
                                                            "last_name": {
                                                                "type": "string",
                                                                "example": "Dupont"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "jean@example.com"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Offre d'emploi introuvable"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/job-applications/{applicationId}/status": {
            "put": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Mettre à jour le statut d'une candidature",
                "description": "Modifie le statut d'une candidature (RH/Admin uniquement)",
                "operationId": "ba8db0678850cfa745ab6fe7da68e6ab",
                "parameters": [
                    {
                        "name": "applicationId",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Nouveau statut de la candidature",
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "interview",
                                            "test",
                                            "accepted"
                                        ],
                                        "example": "interview"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Statut mis à jour avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "interview"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Candidature introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez votre token JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}